Bson的toString输出Date类型错误,在日期为1970年之前的年份 SEQUOIADBMAINSTREAM-1068
git-svn-id: http://192.168.20.11/sequoiadb/trunk@19797 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/bson/bson-inl.h
@@ -713,7 +713,8 @@ namespace bson {
         case bson::Date:
         {
             char buffer[64] ;
-            time_t timer = (time_t)( date() / 1000 ) ;
+            // date() return UINT64, but need INT64
+            time_t timer = (time_t)( ( (long long)date() ) / 1000 ) ;
             struct tm psr ;
             local_time ( &timer, &psr ) ;
             memset ( buffer, 0, 64 ) ;
